流程成功

+++ app/user/user_service.go
@@ -12,7 +12,7 @@ type Service struct {
 func NewService(dao *Dao) *Service {
 	return &Service{dao}
 }
-func (s Service) GetUserById(id int) model.User {
+func (s Service) GetUserById(id int64) model.User {
 	user, err := s.dao.GetUserById(id)
 	if err != nil {
 		panic(errox.WithMessage("用户不存在"))
